[Chore] 필요없는 주석 삭제 및 폴더 정리
