+++ DrawingApp/DrawingApp/Resource/Assets.xcassets/AccentColor.colorset/Contents.json
