+++ DrawingApp/DrawingApp/Resource/Assets.xcassets/AppIcon.appiconset/Contents.json
