+++ DrawingApp/DrawingApp/Resource/Assets.xcassets/Contents.json
